% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_to_points.R
\name{extract2d}
\alias{extract2d}
\title{Extract values from a 2D netCDF variable}
\usage{
extract2d(data, nc, var = NULL, ...)
}
\arguments{
\item{data}{Data frame containing observation points.}

\item{nc}{netCDF source. Can be a file path, vector of file paths, list of file
paths, opened \code{ncdf4} object, list of opened objects, or a data frame with a
file column.}

\item{var}{Name of the variable to extract from the netCDF file. If \code{NULL},
the function tries to detect the variable automatically from each file. This
only works when each netCDF file contains one single variable.}

\item{...}{Additional arguments passed to \code{\link[=extract_netcdf]{extract_netcdf()}}.}
}
\value{
The original \code{data} with extracted values appended.
}
\description{
Use this function when the variable has longitude and latitude dimensions, and
optionally a time dimension, but no depth dimension.
}
