% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_bathymetry}
\alias{load_bathymetry}
\title{Load bathymetry raster}
\usage{
load_bathymetry(file_path)
}
\arguments{
\item{file_path}{Character. Path to bathymetry raster file.}
}
\value{
SpatRaster.
}
\description{
Load a bathymetry raster (e.g., GEBCO).
}
