% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_range_at_depth}
\alias{plot_range_at_depth}
\title{Plot species range at a specific depth layer}
\usage{
plot_range_at_depth(species_range, depth, rast_3d)
}
\arguments{
\item{species_range}{sf or SpatVector. Species range polygon.}

\item{depth}{Numeric. Depth (metres) at which to display environmental data.}

\item{rast_3d}{SpatRaster. Multi-depth raster with layer names following the
\verb{\{variable\}_depth=\{value\}} convention.}
}
\value{
A ggplot object.
}
\description{
Map view of a species range with environmental variable values at a specific
depth layer.
}
