% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_volume_overlap}
\alias{plot_volume_overlap}
\title{Plot 3D volume overlap between two ranges}
\usage{
plot_volume_overlap(overlap_rast)
}
\arguments{
\item{overlap_rast}{SpatRaster. Per-cell overlap raster from
\code{\link[=calc_volume_overlap]{calc_volume_overlap()}}.}
}
\value{
A ggplot object.
}
\description{
Map view of per-cell 3D volume overlap between two rasterized ranges (output
of \code{\link[=calc_volume_overlap]{calc_volume_overlap()}}). Cells colored by overlap depth.
}
