% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woa_volume_extract.R
\name{woa_nc_extract}
\alias{woa_nc_extract}
\title{Extract all layers of a given variable from WOA Raster}
\usage{
woa_nc_extract(woa_nc, selected_field)
}
\arguments{
\item{woa_nc}{SpatRaster. This is the SpatRaster loaded with the terra package.
This is downloaded from the World Ocean Atlas 2023 data page:
https://www.ncei.noaa.gov/access/world-ocean-atlas-2023/}

\item{selected_field}{string. This is the statistical field to select, see
World Ocean Atlas 2023 Product Documentation for further details:
https://repository.library.noaa.gov/view/noaa/70581}
}
\value{
SpatRaster
}
\description{
Extract all layers of a given variable from WOA Raster
}
