% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woa_volume_extract.R
\name{woa_volume_extract}
\alias{woa_volume_extract}
\title{Extract World Ocean Atlas values from 3D volume}
\usage{
woa_volume_extract(area, min_depth, max_depth, woa_nc, selected_field)
}
\arguments{
\item{area}{SpatVector. This is the area to which the WOA raster will get
cropped to.}

\item{min_depth}{numeric. This is the minimum depth in metres of the volume,
which will be rounded to the nearest WOA depth available.}

\item{max_depth}{numeric. This is the maximum depth in metres of the volume,
which will be rounded to the nearest WOA depth available.}

\item{woa_nc}{SpatRaster. This is the SpatRaster loaded with the terra package.
This is downloaded from the World Ocean Atlas 2023 data page:
https://www.ncei.noaa.gov/access/world-ocean-atlas-2023/}

\item{selected_field}{string. This is the statistical field to select, see
World Ocean Atlas 2023 Product Documentation for further details:
https://repository.library.noaa.gov/view/noaa/70581}
}
\value{
SpatRaster
}
\description{
Extract World Ocean Atlas values from 3D volume
}
