% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{calc_volume}
\alias{calc_volume}
\title{Calculate total 3D volume of a rasterized range}
\usage{
calc_volume(range_rast)
}
\arguments{
\item{range_rast}{SpatRaster. Output from \code{\link[=vect_to_envelope]{vect_to_envelope()}}, with layers:
depth_min, depth_max.}
}
\value{
Numeric. Total volume in km³.
}
\description{
Volume = sum of (cell_area x (depth_max - depth_min)) across all present
cells.
}
