% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus.R
\name{copernicus_cache_clear}
\alias{copernicus_cache_clear}
\title{Clear the Copernicus cache}
\usage{
copernicus_cache_clear(confirm = TRUE)
}
\arguments{
\item{confirm}{Logical. Require interactive confirmation. Default \code{TRUE}.}
}
\value{
Invisibly, \code{TRUE} on success.
}
\description{
Remove all files downloaded to the default Copernicus cache.
}
