% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_to_points.R
\name{extract3d_surface}
\alias{extract3d_surface}
\title{Extract the surface layer from a 3D netCDF variable}
\usage{
extract3d_surface(data, nc, var = NULL, ...)
}
\arguments{
\item{data}{Data frame containing observation points.}

\item{nc}{netCDF source. Can be a file path, vector of file paths, list of file
paths, opened \code{ncdf4} object, list of opened objects, or a data frame with a
file column.}

\item{var}{Name of the variable to extract from the netCDF file. If \code{NULL},
the function tries to detect the variable automatically from each file. This
only works when each netCDF file contains one single variable.}
}
\value{
The original \code{data} with extracted values appended.
}
\description{
The surface layer is defined as the first available depth layer in the netCDF
file. This is usually the shallowest layer.
}
