% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{fill_missing_depths}
\alias{fill_missing_depths}
\title{Fill missing depth values}
\usage{
fill_missing_depths(depth_table, method = "genus_mean")
}
\arguments{
\item{depth_table}{Tibble. Output from \code{\link[=fetch_species_depths]{fetch_species_depths()}} with
possible NA values in depth columns.}

\item{method}{Character. Method for filling missing values. Default
\code{"genus_mean"}.}
}
\value{
Complete depth table with no NA depth values.
}
\description{
Fill NA depth values using genus-level means (or other method). Handles
swapped upper/lower values.
}
