% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cumulative_pressure}
\alias{plot_cumulative_pressure}
\title{Plot cumulative fishing pressure on a species}
\usage{
plot_cumulative_pressure(species_rast, fishery_rasters)
}
\arguments{
\item{species_rast}{SpatRaster. Rasterized species range from
\code{\link[=rasterize_range]{rasterize_range()}}.}

\item{fishery_rasters}{List of SpatRasters. Rasterized fishery footprints
from \code{\link[=rasterize_range]{rasterize_range()}}.}
}
\value{
A ggplot object.
}
\description{
Map showing cumulative fishing pressure from all sub-fisheries on a given
species. Cells colored by number of overlapping fisheries.
}
