% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woa.R
\name{woa_cache_clear}
\alias{woa_cache_clear}
\title{Clear the WOA cache}
\usage{
woa_cache_clear(confirm = TRUE)
}
\arguments{
\item{confirm}{Logical. Require interactive confirmation. Default \code{TRUE}.}
}
\value{
Invisibly, \code{TRUE} on success.
}
\description{
Remove all cached WOA NetCDF files.
}
