% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woa.R
\name{woa_nc_extract}
\alias{woa_nc_extract}
\title{Extract all layers of a given variable from a WOA raster}
\usage{
woa_nc_extract(woa_nc, selected_field)
}
\arguments{
\item{woa_nc}{SpatRaster loaded with \code{\link[terra:rast]{terra::rast()}} from a WOA .nc file.
Downloaded from \url{https://www.ncei.noaa.gov/access/world-ocean-atlas-2023/}.}

\item{selected_field}{Character. Statistical field to select. One of:
\code{"an"} (objectively analyzed climatology), \code{"mn"} (statistical mean),
\code{"dd"} (number of observations), \code{"sd"} (standard deviation),
\code{"se"} (standard error), \code{"oa"} (mean minus climatology),
\code{"gp"} (number of mean values within radius of influence),
\code{"sdo"} (objectively analyzed standard deviation),
\code{"sea"} (standard error of the analysis). See the
\href{https://repository.library.noaa.gov/view/noaa/70581}{WOA 2023 Product Documentation}.}
}
\value{
SpatRaster containing only the selected field's depth layers,
using the native \verb{\{variable\}_depth=\{value\}} layer naming convention.
}
\description{
Select layers corresponding to a given WOA statistical field (e.g. \code{"an"},
\code{"mn"}) from a SpatRaster loaded from a WOA NetCDF. Used internally by
\code{\link[=woa_load_nc]{woa_load_nc()}}.
}
